♻️ Refactor user controllers & update image deletion module

+++ backend/modules/delete.image.js
@@ -1,9 +1,11 @@
 const fs = require("fs");
 
 const deleteImage = (imageUrl) => {
-    if (imageUrl) {
-        const filename = imageUrl.split("/images/")[1];
-        fs.unlinkSync(`images/${filename}`);
+    if (imageUrl !== "/images/avatars/default.jpg") {
+        if (imageUrl) {
+            const filename = imageUrl.split("/images/")[1];
+            fs.unlinkSync(`images/${filename}`);
+        }
     }
 };
